+++ frontend/app/src/components/address-book-manager/AddressBookManagement.vue
@@ -13,8 +13,8 @@ import EthNamesHint from '@/components/EthNamesHint.vue';
 import TablePageLayout from '@/components/layout/TablePageLayout.vue';
 import TableFilter from '@/components/table-filter/TableFilter.vue';
 import { type Filters, type Matcher, useAddressBookFilter } from '@/composables/filters/address-book';
-import { useCommonTableProps } from '@/composables/use-common-table-props';
 import { usePaginationFilters } from '@/composables/use-pagination-filter';
+import { useCommonTableProps } from '@/modules/table/use-common-table-props';
 import { useAddressesNamesStore } from '@/store/blockchain/accounts/addresses-names';
 
 const { t } = useI18n();