+++ frontend/app/src/types/settings/frontend-settings.ts
@@ -137,6 +137,7 @@ export const FrontendSettings = z.object({
   minimumDigitToBeAbbreviated: z.number().default(MINIMUM_DIGIT_TO_BE_ABBREVIATED),
   nftsInNetValue: z.boolean().default(true),
   notifyNewNfts: z.boolean().optional().default(false),
+  persistTableSorting: z.boolean().default(false),
   profitLossReportPeriod: ProfitLossTimeframe.default({
     quarter: Quarter.ALL,
     year: new Date().getFullYear().toString(),