+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`8928` Users will now have the option to keep the table sorting preference.
 * :feature:`9495` rotki now supports Stake DAO across all supported EVM chains, including balance tracking.
 * :feature:`-` Now free users can change the theme to dark mode too.
 * :feature:`8012` rotki will now support Pendle Finance, with related historical events properly decoded and correct prices and balances shown.